chore: reused interfaces and added initial state
